Define stack_t in sys/_sigaltstack.h
The sigaltstack(2) definition needs this type so make it available
without all of sys/signal.h.

Reviewed by:	kib
Differential Revision:	https://reviews.freebsd.org/D44383

(cherry picked from commit fba329a195e8cfd52a0c59314d40f0143f32af10)
